Independent coin values
Changed the coin history script to store coin values individually in a new array instead of using the heads or tails value for each coin, this allows Lucky Charm effects to target specific coins without affecting the base wager and thus all matching coins. Also updated the run calculation function to be compatible with this.