Item classifications
Added a typing to items to indicate when they activate, the types being
-Immediate
-Addition
-Post_Run
-Utility
Most items i've added thus far are either Addition (occurring after the wager value is added to coins) and Utility, an item that isnt used in the scoring sequence (ie Spatula or train ticket). The items have a small label on the bottom right to indicate their type for savvy players.