Added Item Swapping
Updated the score calculation script to have a button that swaps the order of items. Only really relevant for one specific item and when you have two items that activate in the same step and want one to go before the other. Would need to be reworked in the circumstance we have more than 2 items at a time.