Pendulum Secret Feature
Added a secret rule to pendulum where if it's your only item, it activates both of it's effects that are dependant on which slot it's in, left or right. since single items are displayed in the centre it should logically get neither or both of the effects, I decided on both, since previously it was defaulting to tails.